[MIM-46] Home, Plan ViewModel 분리
Home, Plan ViewModel 분리

- 한 Fragment에서 이동한 후 다시 돌아갔을때 DispatchedTask에서 JobCancelEception이 발생해서
- viewModel을 분리하고, viewModelScope를 CoroutineScope(Dispatchers.IO)로 변경
- 수정후 정상동작하는데 아직 정확한 원인은 좀 더 공부를 해봐야할것같다
