+++ presentation/src/main/java/com/moon/morningismiracle/home/HomeViewModel.kt
@@ -3,8 +3,14 @@ package com.moon.morningismiracle.home
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.moon.domain.model.PlanModel
+import com.moon.domain.model.PlanState
 import com.moon.domain.usecase.GetPlanListUseCase
+import com.moon.domain.usecase.SetPlanDelayOneDayUseCase
+import com.moon.domain.usecase.SetPlanStateUseCase
+import com.moon.morningismiracle.di.DateInfo
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.collect
@@ -14,17 +20,33 @@ import javax.inject.Inject
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
-    private val getPlanListUseCase: GetPlanListUseCase
+    private val getPlanListUseCase: GetPlanListUseCase,
+    private val setPlanStateUseCase: SetPlanStateUseCase,
+    private val setPlanDelayOneDayUseCase: SetPlanDelayOneDayUseCase,
 ): ViewModel() {
 
     private val _planDataList = MutableStateFlow<List<PlanModel>>(emptyList())
     val planDataList: StateFlow<List<PlanModel>> = _planDataList
 
     fun getPlanList(date: Date) {
-        viewModelScope.launch {
+        CoroutineScope(Dispatchers.IO).launch {
             getPlanListUseCase(date = date).collect { list ->
                 _planDataList.value = list
             }
+        }
+    }
+
+    fun setPlan(planId: Long, state: PlanState) {
+        CoroutineScope(Dispatchers.IO).launch {
+            setPlanStateUseCase(planId, state)
+            getPlanList(DateInfo.today)
+        }
+    }
+
+    fun setPlanDelayOndDay(planId: Long, newDate: Date) {
+        CoroutineScope(Dispatchers.IO).launch {
+            setPlanDelayOneDayUseCase(planId, newDate)
+            getPlanList(DateInfo.today)
         }
     }
 }
