+++ presentation/src/main/java/com/moon/morningismiracle/plan/PlanViewModel.kt
@@ -1,11 +1,8 @@
 package com.moon.morningismiracle.plan
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import com.moon.domain.model.PlanModel
-import com.moon.domain.model.PlanState
 import com.moon.domain.usecase.*
-import com.moon.morningismiracle.di.DateInfo
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -20,40 +17,23 @@ import javax.inject.Inject
 class PlanViewModel @Inject constructor(
     private val addPlanUseCase: AddPlanUseCase,
     private val deletePlanUseCase: DeletePlanUseCase,
-    private val setPlanStateBeforeToday: SetPlanStateBeforeToday,
     private val getPlanListUseCase: GetPlanListUseCase,
-    private val setPlanStateUseCase: SetPlanStateUseCase,
-    private val setPlanDelayOneDayUseCase: SetPlanDelayOneDayUseCase,
 ) : ViewModel() {
-
-    private var _planDataList = MutableStateFlow<List<PlanModel>>(emptyList())
-    val planDataList: StateFlow<List<PlanModel>> = _planDataList
 
     private var _selectDayPlanDataList = MutableStateFlow<List<PlanModel>>(emptyList())
     val selectDayPlanDataList: StateFlow<List<PlanModel>> = _selectDayPlanDataList
 
     fun addPlan(planData: PlanModel) {
-        viewModelScope.launch {
+        CoroutineScope(Dispatchers.IO).launch {
             addPlanUseCase(planData)
+            getSelectDayPlanList(planData.planDate)
         }
     }
 
     fun deletePlan(planData: PlanModel, isSelectedDate: Boolean) {
-        viewModelScope.launch {
-            deletePlanUseCase(planData)
-            if (isSelectedDate) {
-                getSelectDayPlanList(planData.planDate)
-            } else {
-                getPlanList(planData.planDate)
-            }
-        }
-    }
-
-    fun getPlanList(date: Date) {
         CoroutineScope(Dispatchers.IO).launch {
-            getPlanListUseCase(date = date).collect { list ->
-                _planDataList.value = list
-            }
+            deletePlanUseCase(planData)
+            getSelectDayPlanList(planData.planDate)
         }
     }
 
@@ -62,20 +42,6 @@ class PlanViewModel @Inject constructor(
             getPlanListUseCase(date = date).collect { list ->
                 _selectDayPlanDataList.value = list
             }
-        }
-    }
-
-    fun setPlan(planId: Long, state: PlanState) {
-        viewModelScope.launch {
-            setPlanStateUseCase(planId, state)
-            getPlanList(DateInfo.today)
-        }
-    }
-
-    fun setPlanDelayOndDay(planId: Long, newDate: Date) {
-        viewModelScope.launch {
-            setPlanDelayOneDayUseCase(planId, newDate)
-            getPlanList(DateInfo.today)
         }
     }
 }
